requirements.txt
pypdfium2
Pillow
pytesseract
tqdm
argparse
